+++ android/src/main/java/com/bitmovin/player/reactnative/MediaSessionPlaybackManager.kt
@@ -9,41 +9,30 @@ import com.bitmovin.player.api.Player
 import com.bitmovin.player.reactnative.extensions.playerModule
 import com.bitmovin.player.reactnative.services.MediaSessionPlaybackService
 import com.facebook.react.bridge.*
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 
-class MediaSessionConnectionManager(val context: ReactApplicationContext) {
-    private var isServiceStarted = false
+class MediaSessionPlaybackManager(val context: ReactApplicationContext) {
     private lateinit var playerId: NativeId
+    internal var serviceBinder: MediaSessionPlaybackService.ServiceBinder? = null
 
-    private val _serviceBinder = MutableStateFlow<MediaSessionPlaybackService.ServiceBinder?>(null)
-    val serviceBinder = _serviceBinder.asStateFlow()
-
-    inner class MediaSessionServiceConnection : ServiceConnection {
+    inner class MediaSessionPlaybackServiceConnection : ServiceConnection {
         override fun onServiceConnected(className: ComponentName, service: IBinder) {
-            // We've bound to the Service, cast the IBinder and get the Player instance
             val binder = service as MediaSessionPlaybackService.ServiceBinder
-            _serviceBinder.value = binder
+            serviceBinder = binder
             binder.player = getPlayer()
         }
 
         override fun onServiceDisconnected(name: ComponentName) {
-            _serviceBinder.value?.player = null
+            serviceBinder?.player = null
         }
     }
 
-    fun setupMediaSession(playerId: NativeId) {
-        this@MediaSessionConnectionManager.playerId = playerId
+    fun setupMediaSessionPlayback(playerId: NativeId) {
+        this.playerId = playerId
+
         val intent = Intent(context, MediaSessionPlaybackService::class.java)
         intent.action = Intent.ACTION_MEDIA_BUTTON
-
-        val connection = MediaSessionServiceConnection()
+        val connection: ServiceConnection = MediaSessionPlaybackServiceConnection()
         context.bindService(intent, connection, Context.BIND_AUTO_CREATE)
-
-        if (!isServiceStarted) {
-            context.startService(intent)
-            isServiceStarted = true
-        }
     }
 
     private fun getPlayer(