+++ src/playbackConfig.ts
@@ -47,19 +47,22 @@ export interface PlaybackConfig {
    * When set to `true`, also make sure to properly configure your app to allow
    * background playback.
    *
-   * On tvOS, background playback is only supported for audio-only content.
-   *
    * Default is `false`.
+   *
+   * @note
+   * On Android, {@link LockScreenControlConfig.isEnabled} has to be `true` for
+   * background playback to work.
+   * @note
+   * On tvOS, background playback is only supported for audio-only content.
    *
    *  @example
    * ```
    * const player = new Player({
-   *   {
+   *   playbackConfig: {
    *     isBackgroundPlaybackEnabled: true,
-   *   }
-   * })
+   *   },
+   * });
    * ```
-   * @platform iOS, tvOS
    */
   isBackgroundPlaybackEnabled?: boolean;
   /**